fix: 타 파트(Question) ID 타입 불일치 컴파일 에러 해결

+++ backend/src/main/java/com/example/Piroin/project/domain/question/service/QuestionService.java
@@ -171,7 +171,7 @@ public class QuestionService {
     }
 
     private StudySession findSession(Long sessionId) {
-        return curriculumRepository.findById(Math.toIntExact(sessionId))
+        return curriculumRepository.findById(Math.toIntExact(sessionId.intValue()))
                 .orElseThrow(() -> new QuestionException(HttpStatus.NOT_FOUND, "세션을 찾을 수 없습니다."));
     }
